Add test, move settings to pkg, more comments
