+++ pkg/security/security.go
@@ -16,9 +16,12 @@ package security
 
 import (
 	"context"
+	"strings"
 	"time"
 
 	"google.golang.org/grpc"
+
+	"istio.io/pkg/env"
 )
 
 const (
@@ -38,6 +41,23 @@ const (
 	// Credential fetcher type
 	GCE  = "GoogleComputeEngine"
 	Mock = "Mock" // testing only
+)
+
+// TODO: For 1.8, make sure MeshConfig is updated with those settings,
+// they should be dynamic to allow migrations without restart.
+// Both are critical.
+var (
+	// Require 3P TOKEN disables the use of K8S 1P tokens. Note that 1P tokens can be used to request
+	// 3P TOKENS. A 1P token is the token automatically mounted by Kubelet and used for authentication with
+	// the Apiserver.
+	Require3PToken = env.RegisterBoolVar("REQUIRE_3P_TOKEN", false,
+		"Reject k8s default tokens, without audience. If false, default K8S token will be accepted")
+
+	// TokenAudiences specifies a list of audiences for SDS trustworthy JWT. This is to make sure that the CSR requests
+	// contain the JWTs intended for Citadel.
+	TokenAudiences = strings.Split(env.RegisterStringVar("TOKEN_AUDIENCES", "istio-ca",
+		"A list of comma separated audiences to check in the JWT token before issuing a certificate. "+
+			"The token is accepted if it matches with one of the audiences").Get(), ",")
 )
 
 // Options provides all of the configuration parameters for secret discovery service
