+++ security/pkg/server/ca/authenticate/kube_jwt.go
@@ -15,6 +15,8 @@
 package authenticate
 
 import (
+	"encoding/base64"
+	"encoding/json"
 	"fmt"
 	"strings"
 
@@ -22,6 +24,7 @@ import (
 	"google.golang.org/grpc/metadata"
 	"k8s.io/client-go/kubernetes"
 
+	"istio.io/istio/pkg/security"
 	"istio.io/istio/security/pkg/k8s/tokenreview"
 )
 
@@ -30,6 +33,9 @@ const (
 	identityTemplate         = "spiffe://%s/ns/%s/sa/%s"
 	KubeJWTAuthenticatorType = "KubeJWTAuthenticator"
 )
+
+// TODO: move this under pkg/k8s - it depends on k8s client library.
+// The cert and OIDC are independent of k8s.
 
 type RemoteKubeClientGetter func(clusterID string) kubernetes.Interface
 
@@ -48,6 +54,40 @@ type KubeJWTAuthenticator struct {
 }
 
 var _ Authenticator = &KubeJWTAuthenticator{}
+
+type jwtPayload struct {
+	// Aud is JWT token audience - used to identify 3p tokens.
+	// It is empty for the default K8S tokens.
+	Aud []string `json:"aud"`
+}
+
+// isK8SUnbound detects if the token is a K8S unbound token.
+// It is a regular JWT with no audience and expiration, which can
+// be exchanged with bound tokens with audience.
+//
+// This is used to determine if we check audience in the token.
+// Clients should not use unbound tokens except in cases where
+// bound tokens are not possible.
+func isK8SUnbound(jwt string) bool {
+	jwtSplit := strings.Split(jwt, ".")
+	if len(jwtSplit) != 3 {
+		return false // unbound tokens are valid JWT
+	}
+	payload := jwtSplit[1]
+
+	payloadBytes, err := base64.RawStdEncoding.DecodeString(payload)
+	if err != nil {
+		return false // unbound tokens are valid JWT
+	}
+
+	structuredPayload := &jwtPayload{}
+	err = json.Unmarshal(payloadBytes, &structuredPayload)
+	if err != nil {
+		return false
+	}
+
+	return len(structuredPayload.Aud) == 0
+}
 
 // NewKubeJWTAuthenticator creates a new kubeJWTAuthenticator.
 func NewKubeJWTAuthenticator(client kubernetes.Interface, clusterID string,
@@ -80,7 +120,25 @@ func (a *KubeJWTAuthenticator) Authenticate(ctx context.Context) (*Caller, error
 	if kubeClient == nil {
 		return nil, fmt.Errorf("could not get cluster %s's kube client", clusterID)
 	}
-	id, err = tokenreview.ValidateK8sJwt(kubeClient, targetJWT, a.jwtPolicy)
+	var aud []string
+
+	// If the token has audience - we will validate it by setting in in the audiences field,
+	// This happens regardless of Require3PToken setting.
+	//
+	// If 'Require3PToken' is set - we will also set the audiences field, forcing the check.
+	// If Require3P is not set - and token does not have audeince - we will
+	// tolerate the unbound tokens.
+	if !isK8SUnbound(targetJWT) || security.Require3PToken.Get() {
+		aud = security.TokenAudiences
+		// TODO: check the audience from token, no need to call
+		// apiserver if audience is not matching. This may also
+		// handle older apiservers that don't check audience.
+	} else {
+		// No audience will be passed to the check if the token
+		// is unbound and the setting to require bound tokens is off
+		aud = nil
+	}
+	id, err = tokenreview.ValidateK8sJwt(kubeClient, targetJWT, aud)
 	if err != nil {
 		return nil, fmt.Errorf("failed to validate the JWT: %v", err)
 	}
